🔨[fix] Repository 반환값 Optional 로 수정, Service 에서 검증하도록 수정

+++ src/main/java/supercoding/mall/service/ProductService.java
@@ -1,12 +1,15 @@
 package supercoding.mall.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import supercoding.mall.domain.Product;
 import supercoding.mall.domain.AddProductDTO;
 import supercoding.mall.repository.ProductRepository;
+import supercoding.mall.exceptions.NotFoundException;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -26,8 +29,14 @@ public class ProductService {
     }
 
     public void deleteProduct(String productId){
-        productRepository.deleteProduct(productId);
-        serialProductId--;
+        Optional<Product> findProduct = productRepository.findProduct(productId);
+
+        if (findProduct.isPresent()){
+            productRepository.deleteProduct(productId);
+            serialProductId--;
+        }else {
+            throw new NotFoundException("삭제할 상품이 존재하지 않습니다.");
+        }
     }
 
     public List<Product> viewProducts(){
